feat: remove guest sessions, fix ordering, and align DB with frontend
Auth & routing:
- Remove all guest/anonymous session code (signInAsGuest, claimGuestAccount, isGuest)
- Customer routes now allow unrestricted browsing; actions (order, book) redirect to /auth
- Profile page shows sign-in/signup prompt for logged-out users
- CustomerLayout header shows "Sign in" when logged out

Fixes:
- Fix orderService to call 8-arg place_order_and_update_loyalty (removed _login_session mismatch)
- Fix Home.tsx favorites to read correct field names (item_name/item_price not menu_item_*)
- Remove broken session_visit_tracking migration that would conflict with live DB

DB fixes (live):
- Fix get_customer_favorites: item_price INTEGER → NUMERIC to match menu_items.price
- Fix modify_order: now calculates subtotal/tax and inserts order items from new_items
- Fix split_order: now computes per-split amounts and copies all order items

Types:
- Regenerate types.ts from live DB (includes customer_favorites, referrals,
  customer_feedback, split_bills, staff_performance_snapshots, new enums)

+++ supabase/migrations/20260507000008_phase5_customer_features.sql
@@ -314,7 +314,7 @@ RETURNS TABLE (
     menu_item_id UUID,
     item_name TEXT,
     item_description TEXT,
-    item_price INTEGER,
+    item_price NUMERIC,
     item_image_url TEXT,
     category TEXT,
     added_at TIMESTAMPTZ,
